backend/requirements.txt
# Core API
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Machine Learning
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.4
joblib==1.3.2

# Database
motor==3.6.0
pymongo==4.9.2

# Weather API
aiohttp==3.10.11
python-dotenv==1.0.1

# WebSocket
websockets==13.1
